+++ test/src/App.js
@@ -1,6 +1,5 @@
 import { CoreProvider } from "@hackru/frontend-core";
-import Linker from "./Linker";
-import Config from "./Config";
+import { Test } from "./testModule";
 
 function App() {
     let i = 0;
@@ -10,8 +9,40 @@ function App() {
                 i += 1;
                 alert(i);
             }
-        }} Linker={Linker}>
-            {Config}
+        }} Linker={{
+            Test
+        }}>
+            {[
+                {
+                    module: "Test",
+                    params: {
+                        text: "an outer button",
+                        garbage: "YAS"
+                    },
+                    children: [{
+                        module: "Test",
+                        params: {
+                            text: "an inner button"
+                        },
+                        children: [],
+                    }],
+                },
+                {
+                    module: "Test",
+                    params: {
+                        text: "another outer button"
+                    },
+                    children: [],
+                },    
+                {
+                    module: "this one doesn't exist",
+                    params: {
+                        text: "wrapped module",
+                        garbage: "YAS"
+                    },
+                    children: [],
+                }
+            ]}
         </CoreProvider>
     );
 }
